feature/requirements.txt
Flask
sqlalchemy
psycopg2-binary
pandas
pandavro